+++ tiny_wizard/player/player_character.gd
@@ -16,7 +16,7 @@ func hit(damage:=1, from:=Vector2.ZERO):
 
 func _process(delta):
 	if Input.is_action_just_pressed("drop_bomb"):
-		(inventory as QuiverInventory).use_item(self, "Bomb")
+		(inventory as QuiverInventory).use_item(self, "Breach Charge")
 #			inventory.add_to_item('bombs', -1)
 #			var bomb = BOMB_SCENE.instantiate()
 #			bomb.position = position